// src/modules/legacy/tom-server/types.ts
export type Config = MConfig & {
  sender_localpart: string;
  jitsiBaseUrl: string;
  jitsiJwtAlgorithm: string;
  jitsiJwtIssuer: string;
  jitsiJwtSecret: string;
  jitsiPreferredDomain: string;
  jitsiUseJwt: boolean;
  matrix_server: string;
  matrix_internal_host: string;
  matrix_database_host: string;
  oidc_issuer?: string;
  sms_api_key?: string;
  sms_api_login?: string;
  sms_api_url?: string;
  qr_code_url?: string;
  invitation_redirect_url?: string;
  chat_url?: string;
  auth_url?: string;
  matrix_admin_login: string;
  matrix_admin_password: string;
  admin_access_token: string;
  signup_url: string;
  twake_chat: TwakeChatEnvironmentConfig;
  features: {
    common_settings: {
      enabled: boolean;
      application_url: string;
    };
    matrix_profile_updates_allowed: boolean;
    user_profile: {
      default_visibility_settings: {
        visibility: string;
        visible_fields: string[];
      };
    };
    user_directory: {
      enabled: boolean;
    };
    createroom_proxy?: {
      enabled?: boolean;
      on_failure?: {
        max_retries?: number;
        nuke_room?: boolean;
      };
      default_preset?: string;
      encryption?: "allowed" | "enforced" | "disabled";
      is_direct_mask?: DirectChatMask;
      presets?: PresetConfig[];
    };
  };
};
// ...
export interface TwakeChatEnvironmentConfig {
  application_name: string;
  application_welcome_message: string;
  privacy_url: string;
  render_html: boolean;
  hide_redacted_events: boolean;
  hide_unknown_events: boolean;
  issue_id: string;
  registration_url: string;
  twake_workplace_homeserver: string;
  app_grid_dashboard_available: boolean;
  platform: string;
  default_max_upload_avatar_size_in_bytes: string;
  dev_mode: boolean;
  qr_code_download_url: string;
  enable_logs: boolean;
  support_url: string;
  enable_invitations: boolean;
}
// ...
export interface PowerLevelEventContent {
  users_default: number;
  events_default: number;
  invite: number;
  state_default: number;
  kick: number;
  ban: number;
  redact: number;
  notifications: {
    room: number;
  };
  events: {
    "m.reaction": number;
    "m.room.redaction": number;
    "m.room.pinned_events": number;
    "org.matrix.msc3401.call": number;
    "org.matrix.msc3401.call.member": number;
    "m.room.name": number;
    "m.room.topic": number;
    "m.room.avatar": number;
    "m.room.history_visibility": number;
    "m.room.power_levels": number;
    "m.room.encryption": number;
    "m.room.server_acl"?: number;
    "m.room.tombstone"?: number;
  };
  users?: Record<string, number>;
  creator_becomes?: number;
  synapse_preset?: string;
  default_visibility?: "public" | "private";
  allow_is_direct?: boolean;
}

export type DirectChatMask = {
  ban?: number;
  invite?: number;
  kick?: number;
  redact?: number;
  state_default?: number;
  users_default?: number;
  creator_becomes?: number;
  events?: Partial<PowerLevelEventContent["events"]>;
};

export type PresetConfig = PowerLevelEventContent & { name: string };